Extracted WorkItemRepository interface
